requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
mysql-connector-python==8.4.0
